snowflake_dbt_codex/quality/checks.py
import os
import yaml


RULES_PATH = os.getenv("QUALITY_RULES", "./quality/rules/claims.yml")


def load_rules():
    if not os.path.exists(RULES_PATH):
        return {}
    with open(RULES_PATH, "r", encoding="utf-8") as handle:
        return yaml.safe_load(handle)


def run_quality_checks():
    rules = load_rules()
    _ = rules
    # Placeholder for checks: row counts, null thresholds, FK integrity.
    return True
