snowflake_dbt_codex/requirements.txt
apache-airflow==2.9.2
snowflake-connector-python==3.9.0
snowflake-sqlalchemy==1.5.1
kafka-python==2.0.2
dbt-core==1.8.4
dbt-snowflake==1.8.4
pyyaml==6.0.2
pandas==2.2.3
